Fixed linting issues.
