+++ engine/Simulator/simulate_main.py
@@ -1,7 +1,7 @@
-from Components.PrintDetails import print_details
-from Computation.EvaluateStatement import *
+from components.print_details import print_details
+from Computation.evaluate_statement import *
 from Computation.TableMatrix.DisplayMatrix import display_matrix
-from Logging.logging_config import logger
+from engine_logging.logging_config import logger
 from Parser.Checks.CheckBrackets import check_left_and_right_brackets
 from Parser.Checks.CheckForIllegalCharacters import check_for_illegal_characters
 from Parser.Checks.CheckBrackets import UnequalBracketsExcept
